requirements.txt
numpy
pillow
openai
langchain
streamlit
paddleocr


